backend/requirements.txt
pandas
numpy
flask
flask_sqlalchemy
python-dotenv
psycopg2-binary
fastapi
uvicorn
geojson
sqlalchemy
PyJWT
requests

pydantic
pydantic[email]
bcrypt==3.2.0 
passlib==1.7.4
# Authentication and security

PyJWT
python-jose[cryptography]
python-multipart
cryptography

# Database connectivity
asyncpg

# CORS middleware
starlette

# AWS
boto3
